@@ -13,18 +13,20 @@
         "email": "nick.evans@northwestern.edu"
     }],
     "require": {
-        "php": "^7.4|^8",
+        "php": ">=8.1",
         "guzzlehttp/guzzle": "^7.0|^6.0",
-        "northwestern-sysdev/event-hub-php-sdk": "^1.0|^2.0",
+        "northwestern-sysdev/event-hub-php-sdk": "^3.0",
         "laravel/ui": "^3.0|^2.0|^4.0",
         "lcobucci/jwt": "^4.0",
         "socialiteproviders/manager": "~4.0",
         "firebase/php-jwt": "^5.3"
     },
     "require-dev": {
-        "orchestra/testbench": "~6.0",
+        "orchestra/testbench": "~8.0",
         "php-coveralls/php-coveralls": "^2.4",
-        "phpunit/phpunit": "^9.0"
+        "phpunit/phpunit": "^9.0",
+        "laravel/pint": "^1.13",
+        "larastan/larastan": "^2.0"
     },
     "autoload": {
         "psr-4": {
@@ -41,11 +43,6 @@
             "providers": [
                 "Northwestern\\SysDev\\SOA\\Providers\\NuSoaServiceProvider"
             ]
-        }
-    },
-    "config": {
-        "platform": {
-            "php": "7.4"
         }
     }
 }
