@@ -11,7 +11,7 @@ return [
         // The forgerock-direct is for advance use-cases and contingencies.
         'strategy' => env('WEBSSO_STRATEGY', 'apigee'),
         'openAmBaseUrl' => env('WEBSSO_URL_BASE', 'https://uat-nusso.it.northwestern.edu'),
-        
+
         // forgerock-direct & apigee
         'realm' => env('WEBSSO_REALM', 'northwestern'),
         'authTree' => env('WEBSSO_TREE', env('DUO_ENABLED', false) == true ? 'ldap-and-duo' : 'ldap-registry'),
@@ -31,7 +31,7 @@ return [
         // HMAC algorithm we'll register the webhook with -- this must correspond to a type in the EventHub API docs
         'hmacVerificationAlgorithmForRegistration' => env('EVENT_HUB_HMAC_VERIFICATION_ALGORITHM_TYPE_REGISTRATION', 'HmacSHA256'),
 
-         // Matching PHP algorithm type, passed to `hash_hmac()`. You can run `hash_algos()` to see what you have available.
+        // Matching PHP algorithm type, passed to `hash_hmac()`. You can run `hash_algos()` to see what you have available.
         'hmacVerificationAlgorithmForPHPHashHmac' => env('EVENT_HUB_HMAC_VERIFICATION_ALGORITHM_TYPE_PHP', 'sha256'),
     ],
 ];
