@@ -40,7 +40,7 @@ class OpenAM11 implements WebSSOStrategy
         if (config('duo.enabled') === true && $user->getMfaVerified() === false) {
             throw new NoSsoSession($login_url_w_redirect);
         }
-        
+
         return $user->getNetid();
     }
 
@@ -50,7 +50,6 @@ class OpenAM11 implements WebSSOStrategy
         if ($logout_return_to_route !== null) {
             $redirect_to = route($logout_return_to_route, [], false);
         }
-        
 
         return redirect($this->sso->getLogoutUrl($redirect_to));
     }
