@@ -3,14 +3,15 @@
 namespace Northwestern\SysDev\SOA\Console\Commands\EventHub;
 
 use Illuminate\Console\Command;
-use Northwestern\SysDev\SOA\EventHub;
 use Illuminate\Console\ConfirmableTrait;
+use Northwestern\SysDev\SOA\EventHub;
 
 class WebhookToggle extends Command
 {
     use ConfirmableTrait;
 
     protected $signature = 'eventhub:webhook:toggle {status : pause or unpause} {queues?* : one or more queues to toggle. if unspecified, all queues will be updated.}';
+
     protected $description = 'Pause or unpause webhook deliveries.';
 
     protected $webhook_api;
@@ -29,11 +30,12 @@ class WebhookToggle extends Command
         $status = strtolower($this->argument('status'));
         if (in_array($status, ['pause', 'unpause']) === false) {
             $this->error('Invalid status. Please specify pause or unpause.');
+
             return 1;
         }
 
         $specific_queues = $this->argument('queues');
-        if (sizeof($specific_queues) > 0) {
+        if (count($specific_queues) > 0) {
             $queues = $specific_queues;
         } else {
             $hooks = $this->webhook_api->listAll();
@@ -48,6 +50,7 @@ class WebhookToggle extends Command
             } catch (EventHub\Exception\EventHubError $e) {
                 $this->error(vsprintf('Unable to change status of "%s".', [$queue]));
                 $this->line('');
+
                 continue;
             }
         }
