@@ -8,14 +8,15 @@ use Illuminate\Support\Str;
 class ShowOAuthCallbackUrl extends Command
 {
     protected $signature = 'websso:callback';
-    protected $description = "Displays the OAuth callback URI, which must be added to an allow list in Azure AD";
+
+    protected $description = 'Displays the OAuth callback URI, which must be added to an allow list in Azure AD';
 
     public function handle()
     {
         $callbackUrl = route('login-oauth-callback', [], true);
 
         $this->line("You must add the OAuth callback URLs to Azure AD's list of web redirect URIs.");
-        $this->line("Shown below is the callback URI for this application:");
+        $this->line('Shown below is the callback URI for this application:');
         $this->newLine();
 
         $this->info($callbackUrl);
@@ -23,9 +24,9 @@ class ShowOAuthCallbackUrl extends Command
 
         if (
             Str::startsWith($callbackUrl, 'http://')
-            && !Str::contains($callbackUrl, 'localhost')
+            && ! Str::contains($callbackUrl, 'localhost')
         ) {
-            $this->error("WARNING: This callback URI is not served over HTTPS. Azure AD will not accept it.");
+            $this->error('WARNING: This callback URI is not served over HTTPS. Azure AD will not accept it.');
         }
     }
 }
